drops mechanic is in progress.
